code-splitting : code-splitting notify.js in src/App.js
